Added icon display in the SD Browser
Qt offers no support for getting the icon of non-existant files. This
likely has to do with some operating systems requiring file contents to
display icons. This is, for now, solved by creating a temporary empty
file and getting the icon of it.

+++ src/imaging/quantize.cpp
@@ -208,7 +208,7 @@ void Cube::setColormapSize(int colorCount) {
 void Cube::swapColors(uint indexA, uint indexB) {
     // Some safety checks
     if (trueColor || (indexA == indexB)) return;
-    if ((indexA >= colors) || (indexB >= colors)) return;
+    if (((int) indexA >= colors) || ((int) indexB >= colors)) return;
 
     // Swap the entries in the colormap
     std::swap(colormap[indexA], colormap[indexB]);
@@ -282,7 +282,7 @@ uint Cube::findColor(Pixel color, uint startIndex) {
     // If equal, we are done entirely
     int minDist = INT_MAX;
     uint mapIdx = 0;
-    for (uint i = startIndex; i < this->colors; i++) {
+    for (uint i = startIndex; (int) i < this->colors; i++) {
         int dist = calc_distance(this->colormap[i], color);
         if (dist < minDist) {
             minDist = dist;
